postgres/pyproject.toml
[tool.poetry]
name = "postgres"
version = "0.1.0"
description = ""
authors = ["charles-42 <71760090+charles-42@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
